♻️ Refactor iscn functions
